requirements.txt
torch
diffusers
transformers
accelerate
peft
streamlit
pillow
requests
